+++ packages/adapter-sveltekit/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @donkeylabs/adapter-sveltekit
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @donkeylabs/server@2.0.6
+
 ## 2.0.5
 
 ### Patch Changes